requirements.txt
arrow==0.12.1
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
cronex==0.1.2
easydict==1.8
gevent==1.3.6
gipc==0.6.0
greenlet==0.4.15
idna==2.7
inotify==0.2.10
Jinja2==2.10
jsonschema==2.6.0
lockfile==0.12.2
MarkupSafe==1.0
more-itertools==4.3.0
msgpack-python==0.5.6
nose==1.3.7
pluggy==0.7.1
prettytable==0.7.2
py==1.6.0
Pygments==2.2.0
pytest==3.8.2
pytest-cov==2.6.0
python-daemon-3K==1.5.8
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
scalpl==0.2.6
six==1.11.0
urllib3==1.23
wishbone==3.1.3
xmltodict==0.11.0
git+https://github.com/openprocurement/galleon.git#egg=galleon
ujson
deepmerge
